requirements.txt
psycopg2==2.9.3
python-dotenv==0.20.0
PyYAML==6.0
